% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plmm-lasso.R
\name{plmm_lasso}
\alias{plmm_lasso}
\title{Fit a high-dimensional PLMM with lasso penalty}
\usage{
plmm_lasso(
  x,
  y,
  series,
  t,
  name_group_var,
  bases,
  gamma,
  lambda,
  timexgroup,
  criterion,
  cvg_tol = 0.001,
  max_iter = 100
)
}
\arguments{
\item{x}{A matrix of covariates.}

\item{y}{A continuous response variable.}

\item{series}{A vector representing the random intercept.}

\item{t}{A numerical vector representing the time variable.}

\item{name_group_var}{The name of the grouping variable in the \code{x} matrix.}

\item{bases}{A matrix of bases functions.}

\item{gamma}{The regularization parameter for the nonlinear effect of time}

\item{lambda}{The regularization parameter for the fixed effects.}

\item{timexgroup}{Logical indicating whether to use a time-by-group interaction.
If \code{TRUE}, each group in \code{name_group_var} will have its own estimate of the time effect.}

\item{criterion}{The information criterion to be used for model selection. Options are "BIC", "BICC", or "EBIC".}

\item{cvg_tol}{Convergence tolerance for the algorithm.}

\item{max_iter}{Maximum number of iterations allowed for convergence.}
}
\value{
A list containing the following components:
\describe{
\item{lasso_output}{A list with the fitted values for the fixed effect and nonlinear effect. The estimated coeffcients for the fixed effects and nonlinear effect. The indices of the used bases functions.}
\item{se}{Estimated standard deviation of the residuals.}
\item{su}{Estimated standard deviation of the random intercept.}
\item{out_phi}{Data frame containing the estimated individual random intercept.}
\item{ni}{Number of timepoitns per observations.}
\item{hyperparameters}{Data frame with lambda and gamma values.}
\item{converged}{Logical indicating if the algorithm converged.}
\item{crit}{Value of the selected information criterion.}
}
}
\description{
Fits a partial linear mixed effects model (PLMM) with a lasso penalty on the fixed effects
and the coefficient associated with the bases functions via penalized maximum likelihood
using the Expectation-Maximization (EM) algorithm. The bases functions represent a nonlinear effect of time.
}
\examples{
\dontrun{
# Generate example data
set.seed(123)
data_sim = simulate_group_inter(N = 50, n_mvnorm = 3, grouped = TRUE,
                                timepoints = 3:5, nonpara_inter = TRUE,
                                sample_from = seq(0,52,13), cst_ni = FALSE,
                                cos = FALSE, A_vec = c(1, 1.5))
sim = data.sim$sim
x = as.matrix(sim[,-1:-3])
y = sim$y
series = sim$series
t = sim$t
bases = create_bases(t)
lambda <- 0.0046
gamma <- 0.00000001
plmm_lasso(x, y, series, t, name_group_var = "group", bases$bases,
gamma = gamma, lambda = lambda, timexgroup = TRUE,
                       criterion = "BIC")
}

}
