% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-simulations.R
\name{simulate_group_inter}
\alias{simulate_group_inter}
\title{Simulate PLMM}
\usage{
simulate_group_inter(
  N = 50,
  n_mvnorm = 100,
  grouped = TRUE,
  timepoints = 3:5,
  nonpara_inter = TRUE,
  sample_from,
  cst_ni,
  cos = FALSE,
  A_vec = c(1, 1.5)
)
}
\arguments{
\item{N}{Number of subjects.}

\item{n_mvnorm}{Number of covariate generates from a multivariate normal distribution.}

\item{grouped}{Logical indicating whether to include grouping effect.}

\item{timepoints}{Vector specifying timepoints for each subject.}

\item{nonpara_inter}{Logical indicating whether the nonparametric function is specific to each group.}

\item{sample_from}{Vector of time points to sample from.}

\item{cst_ni}{Logical indicating whether to sample time points for each subject or fix them.}

\item{cos}{Logical indicating whether to use cosine function for second group.}

\item{A_vec}{Vector of amplitudes for the nonlinear functions for each group.}
}
\value{
A list with three components:
\item{sim}{Simulated data frame.}
\item{phi}{Individual random intercepts.}
\item{f_val}{Values of the nonlinear functions.}
}
\description{
Simulate a partial linear mixed-effects model.
}
\examples{
simulate_group_inter(N = 50, n_mvnorm = 100, grouped = TRUE,
                     timepoints = 3:5, nonpara_inter = TRUE,
                     sample_from = seq(0, 10, by = 0.1), cst_ni = TRUE, cos = FALSE, A_vec = c(1, 1.5))
}
